v1/requirements.txt
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
dateparser>=1.2.0
parsedatetime>=2.6
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
